fix: change double to float

+++ src/accelerations/accelerations_serial.cpp
@@ -14,10 +14,10 @@ void accelerations(Planet b[])
         for (int j = i + 1; j < NUM_BODIES; ++j)
         {
             vec3 dpos = b[j].pos - b[i].pos;
-            double dist2 = dpos.length_squared() + EPSILON;
-            double dist = sqrt(dist2);
+            float dist2 = dpos.length_squared() + EPSILON;
+            float dist = sqrt(dist2);
 
-            double F = (G * b[i].mass * b[j].mass) / dist2;
+            float F = (G * b[i].mass * b[j].mass) / dist2;
             vec3 force = F * dpos / dist;
             b[i].acc += force / b[i].mass;
             b[j].acc -= force / b[j].mass;
